+++ semantic-conventions-genai/src/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/* eslint-disable no-restricted-syntax --
+ * These re-exports are only of constants, only two-levels deep, and
+ * should not cause problems for tree-shakers.
+ */
+
+export * from './stable_attributes';
+export * from './stable_metrics';
+export * from './stable_events';